move timebar activity graph to its own hook
